+++ sdk/storage/azure-storage-blob/src/main/java/com/azure/storage/blob/options/BlobDownloadContentOptions.java
@@ -7,6 +7,7 @@ import com.azure.core.annotation.Fluent;
 import com.azure.storage.blob.models.BlobRange;
 import com.azure.storage.blob.models.BlobRequestConditions;
 import com.azure.storage.blob.models.DownloadRetryOptions;
+import com.azure.storage.common.DataLocalityEndpoint;
 import com.azure.storage.common.ContentValidationAlgorithm;
 
 /**
@@ -19,7 +20,7 @@ public final class BlobDownloadContentOptions {
     private BlobRequestConditions requestConditions;
     private boolean retrieveContentRangeMd5;
     private ContentValidationAlgorithm contentValidationAlgorithm;
-    private String dataLocalityEndpoint;
+    private DataLocalityEndpoint dataLocalityEndpoint;
 
     /**
      * Creates a new instance of {@link BlobDownloadContentOptions}.
@@ -133,17 +134,17 @@ public final class BlobDownloadContentOptions {
      *
      * @return The data locality endpoint, or null if the client's configured endpoint should be used.
      */
-    public String getDataLocalityEndpoint() {
+    public DataLocalityEndpoint getDataLocalityEndpoint() {
         return dataLocalityEndpoint;
     }
 
     /**
      * Sets the endpoint to use for this download.
      *
-     * @param dataLocalityEndpoint The absolute endpoint selected from blob layout information.
+     * @param dataLocalityEndpoint The endpoint selected from blob layout information.
      * @return The updated options.
      */
-    public BlobDownloadContentOptions setDataLocalityEndpoint(String dataLocalityEndpoint) {
+    public BlobDownloadContentOptions setDataLocalityEndpoint(DataLocalityEndpoint dataLocalityEndpoint) {
         this.dataLocalityEndpoint = dataLocalityEndpoint;
         return this;
     }