+++ sdk/storage/azure-storage-blob/src/main/java/com/azure/storage/blob/specialized/BlobAsyncClientBase.java
@@ -95,6 +95,7 @@ import com.azure.storage.blob.options.BlobSetAccessTierOptions;
 import com.azure.storage.blob.options.BlobSetTagsOptions;
 import com.azure.storage.blob.sas.BlobServiceSasSignatureValues;
 import com.azure.storage.common.ContentValidationAlgorithm;
+import com.azure.storage.common.DataLocalityEndpoint;
 import com.azure.storage.common.StorageSharedKeyCredential;
 import com.azure.storage.common.Utility;
 import com.azure.storage.common.implementation.util.AutoRefreshingCache;
@@ -1637,7 +1638,7 @@ public class BlobAsyncClientBase {
                         BlobLayoutCacheValue::getExpiresOn);
                     AutoRefreshingCache<BlobLayoutCacheValue> finalLayoutCache = layoutCache;
                     chunkDownloadFunc = (range, conditions) -> finalLayoutCache.getValidValueAsync().flatMap(cached -> {
-                        String endpoint
+                        DataLocalityEndpoint endpoint
                             = BlobLayoutRangeResolver.resolveEndpoint(range.getOffset(), cached.getRanges());
                         Context callContext = StorageImplUtils.addDataLocalityEndpoint(finalContext, endpoint);
                         return BlobAsyncClientBase.this.downloadStreamWithResponse(range, downloadRetryOptions,