+++ sdk/storage/azure-storage-file-datalake/CHANGELOG.md
@@ -5,6 +5,8 @@
 ### Features Added
 
 ### Breaking Changes
+- Replaced string-based data locality endpoints in file layout ranges and read options with the validated
+  `DataLocalityEndpoint` type.
 
 ### Bugs Fixed